+++ src/plugin.ts
@@ -7,6 +7,7 @@ import {
 	Rs3GimGroupRankOther
 } from "./actions/rs3-gim-group-rank";
 import { Rs3LevelTracker } from "./actions/rs3-level-tracker";
+import { Rs3WildyEventTimer } from "./actions/rs3-wildy-event-timer";
 
 streamDeck.logger.setLevel(LogLevel.TRACE);
 streamDeck.actions.registerAction(new Rs3GimGroupRankAboveLinked());
@@ -14,4 +15,5 @@ streamDeck.actions.registerAction(new Rs3GimGroupRankBelowLinked());
 streamDeck.actions.registerAction(new Rs3GimGroupRank());
 streamDeck.actions.registerAction(new Rs3GimGroupRankOther());
 streamDeck.actions.registerAction(new Rs3LevelTracker());
+streamDeck.actions.registerAction(new Rs3WildyEventTimer());
 streamDeck.connect();